Analysis renamed for Statistics, implementation generalized and simplified. Non-compiling code moved for Master.py (maybe temporarily).
